Fix allowlist regex in bindgen invocation
The allowlist is just a regex, which means bindgen leaves it to the
user to resolve Windows vs POSIX filepath differences. We need to
support both / and \. It's unclear why only some headers are broken, but
it's probably something to do with whether the header is included
directly or indirectly.

Bug: 595
Change-Id: I8b785aeaaeff162d9eb2aced89928f9602445903