DO NOT SUBMIT YET Switch to bindgen's static inline support
This removes the need to hand-write rust_wrapper.c, because bindgen can
generate it for us. bindgen 0.65 or later is needed. Earlier versions of
this were buggy.

I've also removed the claim that bssl-sys is somehow a solution for
version skew. That was the original thinking from Android, but it hasn't
worked out.

This is not quite yet ready to land. (First I need to replicate it in
Android to make sure it won't break them.)

Fixed: 596
Change-Id: I51fd1108a8c17a06f1bdd9171ebf352cea871723