Remove unimplemented SSL BIO_ctrl values
We have since added an implementation of a subset of the SSL BIO, but we
don't implement all the features, notably some of the BIO_ctrl values.
Remove them, so it doesn't look like they should work.

Update-Note: I found no code using those symbols (that we build). If
anything was, they most likely were broken. Now they'll fail to build
and the brokenness will be more obvious. (If we find something needs it,
we can always go back and implement them.)

Fixed: 420
Change-Id: Iad03fa65f098023dca555a9b2ac0214ba4264546