Don't make assumptions about GCM128_CONTEXT layout in aesni-gcm-x86_64.pl
This is a little trickier because Intel architectures are so
inconveniently register-starved. This will was already using every
register. However, since Xi is only needed at the start and end of the
function, I just swapped the Xi and Htable parameters. Xi is passed on
the stack, so we don't need to explicitly spill it.

Change-Id: I2ef4552fc181a5350c9b1c733cf2319377a06b74