Introduce Zephyr module definition file
ChromeOS Fingerprint MCU[1] (FPMCU for short) uses symmetric and
asymmetric cryptography to encrypt/decrypt templates and authenticate
user.

We already use BoringSSL in CrosEC[2] (which is the RTOS of the FPMCU),
but since we are moving to Zephyr RTOS[3] we need to support BoringSSL
library as well.

Zephyr build system has support for external modules[4], which is
convenient way to integrate well-established code in projects that use
Zephyr.

To recognize external project as a Zephyr module, the module repository
codebase should have 'module.yml' file in 'zephyr' directory.

Minimal module.yml file is as follows:

build:
  cmake-ext: True
  kconfig-ext: True

It indicates that CMakeLists.txt and Kconfig files necessary to build
the project are located externally to the Zephyr module itself (under
the module external root path).

[1] https://chromium.googlesource.com/chromiumos/platform/ec/+/HEAD/docs/fingerprint/fingerprint.md
[2] https://chromium.googlesource.com/chromiumos/platform/ec/+/HEAD/README.md
[3] https://zephyrproject.org/
[4] https://docs.zephyrproject.org/latest/develop/modules.html

Bug: b:281522688
Change-Id: I9010e05f372d7683c0240ab969821cb1ce0c9bcc