Update googletest
It's been a couple years since we did that. Mostly no modifications
needed, except a missing static on CreateArgvFromArgs. We probably could
turn -Wmissing-prototypes off in googletest, but so far it's caught a
an issue worth fixing each time, so just fix it and send the fix
upstream for now.

Change-Id: Ifaf3476f2079d444512a235756cfe54f492b9c92