DO NOT MERGE: Add low level bindings to AES-CTR
Just pushing this as an expeirment for now, since I already had it
implemented. We will likely not want to expose this low level API
as it leaves some performance on the table. Instead we will need
to go through either/both of the EVP_cipher API or AEAD API

Change-Id: I095d7aa2a491e44226c5eabecdbbe4890897b2f4