Add support for fork detection via pthread_atfork()
This provides an alternative way for fork detection that can
be used if we are certain the platform does not have other ways
of cloning an address space in use that are not noticed by
pthread_atfork.

This supports using pthread_atfork for fork detection on
macOS, OpenBSD and FreeBSD. Linux also now uses atfork
in addition to MADVISE, potentially returning a generation
number increased by both methods, or only by pthread_atfork()
if MADVISE is not supported by the kernel.

Change-Id: I79fd7769477dc90bfe37229d2ff2e8c16898dff7