Add support for fork detection via pthread_atfork()
This provides an alternative way for fork detection that can
be used if we are certain the platform does not have other ways
of cloning an address space in use that are not noticed by
pthread_atfork.

This supports using pthread_atfork for fork detection on
macOS, OpenBSD and FreeBSD. Linux also now uses atfork
in addition to MADVISE. I have left it returning 0 for the
moment in the case of a kernel not suportting MADV_WIPEONFORK.
so such a system will still draw entropy on every call.  We
could possibly decide to change that and rely on atfork
in this case if we wanted.

Change-Id: I79fd7769477dc90bfe37229d2ff2e8c16898dff7