Add a value barrier when checking for point doubling.
Many of our point addition functions internally check for the doubling
case and branch because the addition formulas are incomplete. This
branch is fine because the multiplication formulas are arranged to not
hit this case.

However, we don't want to leak the couple of intermedate values that
determine whether to branch. Previously, we ran into this with
https://boringssl-review.googlesource.com/c/boringssl/+/36465.

Moritz Schneider, Nicolas Dutly, Daniele Lain, Ivan Puddu, and Srdjan
Capkun reported that we still leak some of the intermediate values. The
compiler understands constructions like if (a | b) enough to compile
them into two branches.

Fix this by adding a value barrier on the final value. As we're also
intentionally leaking the result of otherwise secret data flow, I've
used the constant_time_declassify functions, which feed into our
valgrind-based constant-time validation and double as barriers.

Accordingly, I've also added some CONSTTIME_SECRET markers around the
ECDSA nonce value, so we can check with valgrind the fix worked. The
marker really should be at a lower level, at ec_random_nonzero_scalar or
further (maybe RAND_bytes?), but for now I've just marked the nonce.

Valgrind now says this code is constant-time, at least up to compilation
differences introduced by the annotations. I've also inspected the
compiler output. This seems to be fine, though neither test is quite
satisfying. Ideally we could add annotations in ways that don't
influence compiler output.

Change-Id: Idfc413a75d92514717520404a0f5424903cb4453