Keep better track of ctr64 vs ctr32 with ChaCha20
Although the assembly function is called ChaCha20_ctr32, the aarch64
implementation actually impelmented ChaCha20_ctr64. This meant that
CRYPTO_chacha_20 gave different results on aarch64 vs. all the other
platforms.

This comes from it being easier to increment a 64-bit general-purpose
register on 64-bit platforms, than to increment just the bottom half.
(Both x86_64 and aarch64 implicitly zero the upper half of the register
when using a 32-bit destination register.)

As a result, the aarch64 assembly file actually implements
ChaCha20_ctr64. (The x86_64 file uses vector registers, so it's ctr32.)
Rename the function accordingly and normalize the behavior in C.

This is purely a theoretical concern. Confidentiality breaks if the
counter wraps to an already-used value, so uses should start at a low
initial value and set input limits to prevent overflow, as
ChaCha20-Poly1305 does. Wraparound is only safe if starting at a high
initial value, then wrapping around to unused lower values, but there is
no reason to offset the counter like this.

Nonetheless, we prefer our functions behave the same across platforms, so
we normalize the behavior.

Fixed: 614
Change-Id: I5e1f50f33b4e31a3d3e0489c84d8563fa51193c5