Fix the combined asm source lists in generate_build_files.py
This fixes one of the two build issues triggered by the curve25519
assembly. The new curve25519 assembly is one file that works for both
ELF and Apple targets. This is handy, but breaks the naive concatenation
logic for making the generated file lists. Fix the generation to
deduplicate first.

Convincing Starlark in Bazel to deduplicate is annoyingly difficult, so
we'll just generate both lists. This results in some duplication in the
generated files, but they're generated so this is fine. Hopefully we
can, in time, just remove the per-platform lists when everyone is
migrated over. To that end, this CL adds the combined output to GN so I
can migrate Chromium.

This CL is not sufficient to unbreak the Bazel build. The next change is
also needed.

Bug: 542
Change-Id: Ibe76ff557fc43f7b4d984ccdf298f13c20f3b50c